+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2025.09.30',
+  '2025.10.07',
   // Fallback runtime versions.
+  '2025.09.30',
   '2025.09.22',
-  '2025.09.16',
 ];
 
 /// Sets the current runtime versions.
@@ -62,7 +62,7 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .github/workflows/all-test.yml and Dockerfile.app
 final String runtimeSdkVersion = '3.9.0';
-final String toolStableDartSdkVersion = '3.9.3';
+final String toolStableDartSdkVersion = '3.9.4';
 final String toolStableFlutterSdkVersion = '3.35.5';
 
 final semanticToolStableDartSdkVersion = Version.parse(